separated NormalizedCompressionMarkovTable
