+++ Makefile
@@ -10,6 +10,7 @@ objects = src/turingMachine.o \
 		  src/parseArgs.o \
 		  src/util.o \
 		  src/tm.o \
+		  src/NormalizedCompressionMarkovTable.o \
 		  src/main.o
 
 %.o: %.cpp
@@ -29,9 +30,11 @@ main.o: src/main.cpp src/tm.h src/parseArgs.h
 
 src/metrics.o: src/metrics.cpp src/metrics.h src/turingMachine.h
 
-src/tm.o: src/tm.cpp src/tm.h src/turingMachine.h src/markovTable.h src/util.h
+src/tm.o: src/tm.cpp src/tm.h src/turingMachine.h src/markovTable.h src/util.h src/NormalizedCompressionMarkovTable.h
 
 src/markovTable.o: src/markovTable.cpp src/markovTable.h src/turingMachine.h src/util.h src/metrics.h
+
+src/NormalizedCompressionMarkovTable.o: src/markovTable.h src/metrics.h
 
 src/parseArgs.o: src/parseArgs.cpp src/parseArgs.h src/traversal.h src/util.h
